Update test/check dependencies and fix new PMD/Checkstyle violations

+++ src/main/java/de/hhu/stups/prob/translator/BNumber.java
@@ -55,11 +55,11 @@ public class BNumber extends Number implements BValue {
 
     @Override
     public float floatValue() {
-        return this.value.floatValue();
+        return this.value;
     }
 
     @Override
     public double doubleValue() {
-        return this.value.doubleValue();
+        return this.value;
     }
 }
